Improved the handling of SNI, reworked the API slightly.
UserData => map[string]string .. to make it more useful directly.
UserObjects => is the old UserData, with map[string]interface{}

Added SetConnectScheme(), to give better control as to how to interprete
a mitm'd CONNECT session.

Reworked docs.

Now ctx.Host() always has a port, and its infered automatically by the
scheme on the first request, if not present.


First draft of a redirectable RoundTripper, where we can connect to a
different host/ip, but still fake that we are talking to the right server.